+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 23.8.1
+- Expanded feedback widget functionality. Added ability to use rating widgets.
+- Added functionality to access tags for feedback widgets.
+
 ## 23.8.0
 - Added `CountlyViewTracking:` interface with new view methods:
     - `setGlobalViewSegmentation:`